🚚 rename: isValidMemberId를 utils로 이동 #67
