+++ src/common/components/atoms/AppLogo.tsx
@@ -1,7 +1,7 @@
 import { Link } from "react-router-dom";
 import Logo from "@/assets/logo.svg?react";
 import useAuthStore from "@/common/stores/authStore";
-import isValidMemberId from "@/common/isValidMemberId";
+import isValidMemberId from "@/common/utils/isValidMemberId";
 
 export default function AppLogo() {
   const { memberId } = useAuthStore((state) => ({
